Fix symlinks

+++ symlinks.sh
@@ -9,6 +9,8 @@ NESTED=$(find "$REPO_PATH" -mindepth 2 -type f \( -iname '*' ! -ipath '*.git/*' 
 
 while read -r broken_link
 do
+    [[ -L "$broken_link" ]] || continue
+    [[ -e "$broken_link" ]] && continue
     target=$(readlink "$broken_link")
     [[ "$target" == "${HOME}/.dotfiles/"* ]] && unlink "$broken_link"
 done <<< "$(find "${HOME}" -maxdepth 3 -type l ! -exec test -e {} \; -print)"
